pyproject.toml
[project]
name = "MaaXuexi"
version = "0.1.0"
description = "基于全新架构的 学习强国 自动化助手。图像技术 + 模拟控制，解放双手！由 MaaFramework 强力驱动！"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.8",
    "httpx>=0.28.1",
    "maafw>=3.0.3",
    "onnxruntime>=1.20.1",
    "opencv-python-headless>=4.11.0.86",
    "plyer>=2.1.0",
    "uvicorn>=0.34.0",
]
[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
